requirements.txt
APScheduler==3.5.1
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
idna==2.6
pkg-resources==0.0.0
pyasn1==0.4.2
pytz==2018.4
PyYAML==3.12
requests==2.18.4
rsa==3.4.2
six==1.11.0
tzlocal==1.5.1
urllib3==1.22
